+++ src/screens/Assets/Assets.js
@@ -25,7 +25,7 @@ import { useTranslation } from 'translations/translate';
 // Components
 import { Container } from 'components/layout/Layout';
 import HeaderBlock from 'components/HeaderBlock';
-import TabView from 'components/modern/TabView';
+import TabView from 'components/layout/TabView';
 
 // Constants
 import { ASSET_CATEGORY as CATEGORY } from 'constants/assetsConstants';
